third_party/blink/web_tests/flag-specific/disable-layout-ng/accessibility/table-with-presentation-role-expected.txt

This is a testharness.js-based test.
FAIL ARIA presentation role on table element has no rows or cells. assert_equals: expected "AXRole: AXGenericContainer" but got "AXRole: AXStaticText"
Harness: the test ran to completion.
